Addedd all features

+++ p1_games.py
@@ -1,8 +1,8 @@
 class Game:
     def __init__(self, id, title, price):
-        self.id = id
+        self.id = int(id)
         self.title = title
-        self.price = int(price)
+        self.price = float(price)
 
     def str(self):
         return f"{id}) {self.title}     Price: {self.price}"
